+++ src/App.tsx
@@ -5,11 +5,14 @@ import { GlobolStyled } from './styles/global'
 import { BrowserRouter } from 'react-router-dom'
 
 import { defaultTheme } from './styles/themes/default'
+import { CycleContextProvider } from './contexts/CyclesContext'
 export function App() {
   return (
     <ThemeProvider theme={defaultTheme}>
       <BrowserRouter>
-        <Router />
+        <CycleContextProvider>
+          <Router />
+        </CycleContextProvider>
       </BrowserRouter>
       <GlobolStyled />
     </ThemeProvider>
