+++ src/page/History/index.tsx
@@ -1,9 +1,15 @@
+import { useContext } from 'react'
+import { CycleContext } from '../../contexts/CyclesContext'
 import { HistoryContainer, HistoryList, Status } from './styles'
 
 export function History() {
+  const { cycle } = useContext(CycleContext)
+
   return (
     <HistoryContainer>
       <h1>Meu Historico</h1>
+      <pre>{JSON.stringify(cycle, null, 2)}</pre>
+
       <HistoryList>
         <table>
           <thead>
