+++ src/page/Home/components/NewCycleForm/index.tsx
@@ -1,6 +1,6 @@
 import { useContext } from 'react'
 import { useFormContext } from 'react-hook-form'
-import { CycleContext } from '../..'
+import { CycleContext } from '../../../../contexts/CyclesContext'
 import { FormContainer, MinutAmountInput, TaskInput } from './styles'
 
 export function NewCycleForm() {
